Deal with metadataheader not being present
In B44H0400001 no metadata and no metadata header is present, I got a keyerror exception, because the code tries to pop a key that would normally be in the metadata (it tried to pop 'startdatum'). 

In the proposed change is a way to deal with it, although there are many (maybe more chique) ways, so feel free to do it otherwise.