Sort observations if index is not monotonic increasing
For some BRO wells (e.g. GMW000000028739_1), the index (datetime) values are not strictly monotonic. This results in unexpected behavior such as a straight-line, from the starting to the end point, in time series plots and NaN values when performing basic statistics (mean, min, max, etc.).
I applied a fix in the code, but I can imagine you would mabye want to add an optional argument to the from_bro method to handle this case more explicitly.